Documents/ML PROJECT/trustlens/backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
torch
transformers
spacy
pydantic
